challenges/Day 5 Task/razorpay-agent/requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
livekit-murf
python-dotenv